+++ Project Gilbert 2024/Assets/Audio/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a835cd7b0819d1f40abb5f792e8d42f1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
