+++ Project Gilbert 2024/Assets/Audio/Script/Audio.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 278d892ff934db94982cd91c7fe2e932
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
